refactor: change history page
