+++ lib/features/workouts_history/widgets/widgets.dart
@@ -1,2 +1,1 @@
-export 'workout_history_body.dart';
 export 'workouts_history_body.dart';
