+++ lib/features/workouts_history/workouts_history.dart
@@ -1,4 +1,3 @@
 export 'cubit/cubit.dart';
-export 'view/workout_history_page.dart';
 export 'view/workouts_history_page.dart';
 export 'widgets/widgets.dart';
